[diffuse] simulate diffuse material.
The object absorbs half the light that hits it, and reflects the rest
back to the world. The maximal number of bounces is limited to 50 per
sample.

An interesting observation - you can actually see that the first
scanlines are faster because the top of the image is only sky. That
means those rays have no bounce.

spp = num of samples per pixel
bpr = max number of bounces per ray

The number of rays per pixel are roughly:
no hit -> max rays = spp        = 100
hit    -> max rays = spp * bpr  = 500