[diffuse] fix shadow acne.
This is done by discarding hits that are a little too close to the
object.  This avoids the cases where the hit point a little below the
object surface and the ray is shot from within the object are trapped
inside the object, which means they turn out black.

This also improves performance (by a factor of 10!!), because those
rays were very costly - they always reached the max depth.