[diffuse] true lambertian reflection.
This is achieved by picking a random point inside the sphere and
normalizing it onto the surface of that sphere. Since the sphere
is the unit sphere, with a radius of 1, we can do that by scaling
the vector to have a length of 1,

Note that this makes the image a little lighter.