[dielectric] use schlick's approximation to deal with reflection.
The approximation calculates a ratio of reflected rays to refracted rays
depending on the ray's angle with the surface normal. As the angle
increases, so does the chance of a ray to be reflected.

The approximation results in a R [0,1]. We then generate a random
double. If it is lower than R, we reflect. If it is higher, we refract.