[texture] add basic support for textures.
Implements uv mapping for a sphere. The supported textures are:
- Solid color
- Checker pattern

Notice that I think there's an error in the checker pattern. It uses the
point of intersection in the global coordinate system, instead of the
one local to the object that was hit. This will cause the texture to
shift when the object is translated. Ideally, the texture should stay
the same.

I implemented a checker pattern that is based on the uv coordinates,
which are normalized to [0, 1] for each object. This results in a
beach-ball-like appearance which I think is more appealing. I will
upload this in a separate commit as an option, but won't use it yet
because the guide might be heading in a different direction.