[deps] add third-party image header.
I use stb_image because it's only one header and should be easy to use.
And that's what the guide recommends.

I can use some sort of package manager (conan with cmake integration, or
cmake alone) but that seems a little overkill for just reading an image.

Since this is for learning purposes, I'll cut myself some slack ;)