[texture] add modular texture modifiers.
The modifiers are easy to create (see documentation in
src/texture/modifiers.h) and can be combined in any way you like. You
just need to create any base texture and chain modifiers to it.

I used the decorator design pattern to do so. This current modifier only
affects the uv coordinates before they are looked-up in the texture. It
is possible to create many other decorators, f.e. one that tints the
color value with a desired hue, after it was retrieved from the texture.

Only a small number of uv modifiers were implemented so far, and they
are:
- offset_u
- offset_v
- flip_u
- flip_v